fix(cubism): key animation copy ownership on stable boundaries
Animation attributes were keyed to a per-active()-view wrapper owner, so
a source from a second facade view over the unchanged backend was wrongly
rejected. Key the facade animation graph on a per-facade token and let
the session layer enforce the real generation boundary. Also reject copy
sources owned by a different DynamicCubismModelAccess even when the
captured generation numbers coincide, and evaluate bezier handle affine
scaling in double so large non-float origins do not corrupt handle times.

+++ runtime/src/main/java/dev/turboism/adapter/cubism/editor/EditorAnimationTimelineAccess.java
@@ -1460,7 +1460,7 @@ final class EditorAnimationTimelineAccess {
 
         HandleData scaled(final int originFrame, final double factor) {
             return new HandleData(
-                (float) (originFrame + (frame - originFrame) * factor),
+                (float) (originFrame + ((double) frame - originFrame) * factor),
                 value,
                 corner
             );
